+++ AirCasting/Utils/SystemSettings.swift
@@ -1,0 +1,23 @@
+// Created by Lunar on 15/07/2021.
+//
+
+import Foundation
+import SwiftUI
+
+func goToLocationAuthSettings() {
+    if let url = URL(string: UIApplication.openSettingsURLString) {
+        let app = UIApplication.shared
+        if app.canOpenURL(url) {
+            app.open(url, options: [:], completionHandler: nil)
+        }
+    }
+}
+
+func goToBluetoothAuthSettings() {
+    if let url = URL(string: UIApplication.openSettingsURLString) {
+        let app = UIApplication.shared
+        if app.canOpenURL(url) {
+            app.open(url, options: [:], completionHandler: nil)
+        }
+    }
+}
